requirements.txt
streamlit
pytesseract
PyMuPDF
Pillow

